Inclusao de fatura

+++ Payment/src/main/java/br/com/itau/payment/repository/PaymentRepository.java
@@ -10,5 +10,5 @@ public interface PaymentRepository extends CrudRepository<Payment, Long>{
 	Iterable<Payment> findByCreditCardId(Long creditCardId);
 
 	@Transactional
-	Iterable<Payment> deleteByCreditCardId(int creditCardId);
+	Iterable<Payment> deleteByCreditCardId(Long creditCardId);
 }
